requirements.txt
json
os
pathlib
azure-core>=1.28.0
azure-identity==1.14.0
azure-search-documents==11.10.0

langchain-openai==0.0.269
langchain-community==0.0.90

openai==0.28.0
streamlit==1.30.0
requests==2.32.4
python-dotenv==1.0.0
azure-functions==1.23.0
